avoid shell expansion during crm configure (bsc#922751)
If we pass cluster configuration directly to a string-based invocation
of 'crm configure' then we face issues with shell expansion since the
command will be run by Chef's execute resource which will run it within
a shell, e.g. effectively invoking a command like

  sh -c 'crm configure primitive foo params password="foo$bar"'

which would incorrectly attempt to expand '$bar' as a shell parameter,
resulting in most likely setting the password parameter value to 'foo'
if the 'bar' shell variable was not set.

Instead it's safer to pass the configuration as an unexpanded string to
crmsh's STDIN, just like we already do when updating resources.  This
effectively invokes a command like:

  echo 'primitive foo params password="foo$bar"' |
    crm configure load update -

Fixes https://bugzilla.suse.com/show_bug.cgi?id=922751