remove default ring from corosync cookbook
Chef merges default attributes which are Arrays with their override
counterparts, so specifying a default ring means that it will appear
in the node attributes alongside whatever ring is configured in the
proposal.  We don't ever want this default ring, especially since it
only has one member, which causes an error about a mismatch between
the number of members in each ring.  Even if that wasn't an issue, the
IP refers to a whole network rather than a single node, so that would
also cause an error.

In Chef 11 it seems that the Array merging behaviour may have changed:

    https://coderanger.net/arrays-and-chef/

but in any case this merging has been proven to be a problem with the
current codebase, so remove the default ring.